// x-pack/platform/plugins/private/reporting/server/types.d.ts
export type ScheduledReportType = Omit<RawScheduledReport, 'schedule'> & {
    schedule: RruleSchedule;
};
export interface ListScheduledReportApiJSON {
    id: string;
    created_at: RawScheduledReport['createdAt'];
    created_by: RawScheduledReport['createdBy'];
    enabled: RawScheduledReport['enabled'];
    jobtype: RawScheduledReport['jobType'];
    last_run: string | undefined;
    next_run: string | undefined;
    notification: RawScheduledReport['notification'];
    payload?: ReportApiJSON['payload'];
    schedule: RruleSchedule;
    space_id: string;
    title: RawScheduledReport['title'];
}
export type ScheduledReportApiJson = ListScheduledReportApiJSON;